requirements.txt
asgiref==3.8.1
Django==4.2.21
sqlparse==0.5.3
typing_extensions==4.13.0
google-generativeai==0.3.2
python-dotenv==1.0.1
hyperon==0.2.5
